image-generator/requirements.txt
Pillow==10.0.0
python-dotenv==1.0.0
customtkinter==5.1.3
requests==2.31.0